make method called by button return True
The method auto_rfq_from_suppliers returned a list of ids of created
RFQs. This list was used by the tests only. In production a method
called by a button should just return True in case of success.

Now the method returns True, and tests are adapted to figure out the
created rfqs from requisition.purchase_ids instead of using the return
value.