revert the framework to use product.product
for openerp 7.0, the framework agreement was attached to a
product.product. While porting to odoo 8.0, initially it was changed to
be attached to a product.template instead.

After more thinking, it turns out that an agreement should indeed be
associated to a specific variant, because variants can have different
prices. In the demo data, we want an agreement on the ipad (32GB), not
on a generic ipad.