Create change
GitHub Pull Request: https://api.github.com/repos/OCA/purchase-workflow/pulls/129

[ADD] procurement_group_purchase_orders

This module is a backport from version 8.0 to 7.0 of the purchase order

creation from procurement orders. At the time of running the scheduler the

application will search for an existing PO in draft state for the same

supplier and with the same destination location. If found, it will extend

the PO to include the new product and quantity to be procured.



The main difference with version 8.0 is that the application will not

attempt to increase the quantity of a line for an existing PO, because each

line in a PO is associated in 7.0 to the procurement's stock move.



Regards,

Jordi.

Patch-set: 1
Change-id: I17f2f5f620f28331c0b506bd8da6eec3cd20b5fd
Subject: [ADD] procurement_group_purchase_orders
Branch: refs/heads/7.0
Commit: 9002440a6dd78d0188e429e0da0524817a549985
Groups: 9002440a6dd78d0188e429e0da0524817a549985