This PR contains the following fixes / improvements:
- When the RFQ was created from a purchase request line,
if the product has a minimum defined in the product.supplierinfo,
a RFQ line was created with quantity ordered = minimum.
If a second request line was later added, and the sum of
the request lines was < RFQ line quantity, no link was
established between the RFQ line and the new request line.
- Tests are added to include the scenario of creating an
RFQ for a purchase request that contains a product that has minimums.