Fix bug in the logic of the module
If the creation of purchase order lines is made with form view rather than tree view then the module misbehaves, ie, creates a line with sequence 9999. This fixes that without breaking the tree view usage. It also makes sure that the sequence is kept when duplicating a purchase order, which makes more sense. It seems that this was a typo/bug on the original module which was not manifestating when using tree view but becomes clear when allowing editing of the purchase order line with form view.