purchase_request: Speed up stock_move company constraint
Stock move is a table that could increase a lot.
So, it's important avoid slow methods in the middle

The constraint method `stock_move._check_company_purchase_request`
compares `purchase_request_allocation.company_id` vs `stock_move.company_id`
to be sure they have the same value

`pra.company_id` was a related field to
`purchase_request_line_id.request_id.company_id`
without `store=True`
Then this constraint consume a lot of time getting all these iteration of records.
For that it was changed to `store=True` and an index was added to get it faster.

And the method was using a loop getting company record by record.
It is so slow too.
It was changed directly using a query since that the ORM doesn't support `JOIN`
And was added an index to `pra.stock_move_id` field to get `JOIN` faster