[13.0][FIX] purchase_delivery_split pickings same date
When there is multiple pickings with the same planned date on entering
the function `_check_split_pickings` the moves are not grouped in one
picking.
This can happen if the picking planned date is changed by another module
on the call to super in the write method.

This resolve that issue.