[IMP] purchase_discount: flag discount hack
As Odoo isn't using its hook method yet, we still need to be using the
price unit hack that makes two writes in order to be able to correctly
compute the discount. This writes can confound other modules and even
provoke an infinite loop in some situations. We should at least flag it
to allow to ignore these writes.

TT36816