[FIX] purchase_propagate_qty: Ci error due to upstream change
Since https://github.com/odoo/odoo/commit/ba4111e7a9202d801e3d51b55cad9659877e514a,
tests are failing.

Analyzing the roout cause, this module was testing things that are not
done by it, but a side effect of the previous status quo.

As such, we remove the tests that were testing this expected side
effects, as well as removing the reference in the README. If the
behavior wants to be restored, it should be explicitly handled in a
module with its own tests.