purchase_request: fix done purchase_request_line
A purchase request can be set to done even if it hasn't been processed.
Meaning that there's no purchase order for any of the request lines
move_dest_ids, that might be waiting.
In the end, this is an incomplete process, and the related moves will
hang forever for a purchase process that will never end.