[IMP] purchase_advance_payment: Relate payments to purchase
In the module purchase_advance_payment was added the relation of Orders
and Payments, in the Order you can make a Payment, while in the Invoice
too, but the payments relate to the order only if it was made in the
Order, not in the Invoice.

Minor change to the view list in purchase orders, added the new field
payments_date to see the dates of the payments.