[IMP] purchase_delivery_split_date: Split improvements
In order to not browsing the purchase order when splitting, do the split
directly on purchase order lines as we have the modified recordset (for
date_planned field).

Use partition() function in order to group lines per date then per picking
to improve readability.

Use the standard move picking assignation (_assign_picking()) and add the
grouping date in the domain.