[FIX] purchase_order_type: preserve order type when matches company
Before this fix, when onchange event for company_id was fired, type for
purchase order was always set, then sometimes overwritten, even if
former order type was compatible with the new company.

This cause an incompatibility with e.g. purchase_order_type_dashboard.
When PO creation form is accesed coming from PO dashboard, selected
order type was actually igored, because onchange for company was fired
and order type overwritten. This fix it.