[FIX] purchase_request: fix migration script for enterprise migration
If you are migrating using the enterprise service, the table ir_property is already converted and deleted. Also the conversion to company_depenent is done for all the fields in that service too. So when trying to access the table in the openupgrade process it raises an exception. So I'm adding a condition to ensure this table still exists and being able to do the migration with both services!